views - combine remix/restore into ButtonIcon
