+++ src/assets/images/Icons.svelte
@@ -1,0 +1,29 @@
+<script>
+  export let name;
+
+  const icons = [
+    {
+      box: {
+        width: 12,
+        height: 12,
+      },
+      name: 'remix',
+      svg: '<path d="M11.8359 9.37499L9.97969 11.0836C9.62578 11.4375 9 11.1867 9 10.6851V9.74765H8.27344C8.19609 9.74765 8.12109 9.71484 8.06719 9.65624L5.9625 7.37812L6.85547 6.41015L8.72578 8.4328H9V7.3078C9 6.80624 9.62578 6.55546 9.97969 6.90937L11.8359 8.57343C12.0539 8.79843 12.0539 9.15468 11.8359 9.37499ZM0.28125 3.56249H2.43281L4.30313 5.58515L5.19609 4.61718L3.09141 2.33905C3.0375 2.28046 2.9625 2.24765 2.88516 2.24765H0.28125C0.126563 2.24765 0 2.37421 0 2.5289V3.2789C0 3.43593 0.126563 3.56249 0.28125 3.56249ZM11.8359 2.58046L9.97969 0.916398C9.62578 0.562492 9 0.813273 9 1.31249V2.24999H8.27344C8.19609 2.24999 8.12109 2.2828 8.06719 2.3414L2.43281 8.43749H0.28125C0.126563 8.43749 0 8.56405 0 8.71874V9.46874C0 9.62343 0.126563 9.74999 0.28125 9.74999H2.88281C2.96016 9.74999 3.03516 9.71718 3.08906 9.65859L8.72344 3.56249H9V4.68749C9 5.18905 9.62578 5.43984 9.97969 5.08593L11.8359 3.37499C12.0539 3.15468 12.0539 2.79843 11.8359 2.58046Z"/>',
+    },
+    {
+      box: {
+        width: 12,
+        height: 12,
+      },
+      name: 'restore',
+      svg: '<path d="M0.495605 0.495605L1.75766 1.75766C2.84335 0.671782 4.3431 0 5.99998 0C9.30629 0 12.0024 2.69818 12 6.00447C11.9976 9.31614 9.31222 12 6 12C4.4534 12 3.0435 11.4148 1.97956 10.4538C1.85586 10.3421 1.85006 10.1499 1.9679 10.0321L2.44544 9.55454C2.55426 9.44572 2.72978 9.43877 2.84465 9.54118C3.68257 10.2882 4.78778 10.7419 6 10.7419C8.62067 10.7419 10.7419 8.62108 10.7419 6C10.7419 3.37933 8.62108 1.25806 6 1.25806C4.6896 1.25806 3.50427 1.78846 2.64644 2.64644L3.95598 3.95601C4.13889 4.13891 4.00935 4.45161 3.7507 4.45161H0.290323C0.129992 4.45161 0 4.32162 0 4.16129V0.700911C0 0.442258 0.312726 0.312726 0.495605 0.495605Z"/>',
+    },
+  ];
+
+  const displayIcon = (currentName) => icons.find((icon) => icon.name === currentName);
+</script>
+
+<svg
+  viewBox="0 0 {displayIcon(name).box.width} {displayIcon(name).box.height}">
+  {@html displayIcon(name).svg}
+</svg>
