--- src/views/forms-controls/ButtonRemix.svelte
@@ -1,27 +1,0 @@
-<script>
-  import { createEventDispatcher } from 'svelte';
-
-  export let action = 'clicked';
-  export let disabled = false;
-
-  const dispatch = createEventDispatcher();
-</script>
-
-<style>
-  /* components/icon-button > @button-icon */
-</style>
-
-<button
-  on:click={() => dispatch('handleAction', action)}
-  class={`item-toggle action-${action}`}
-  disabled={disabled}
->
-  <span class="label">
-    Remix text
-  </span>
-  <span class="icon">
-    <svg viewBox="0 0 12 12">
-      <path d="M11.8359 9.37499L9.97969 11.0836C9.62578 11.4375 9 11.1867 9 10.6851V9.74765H8.27344C8.19609 9.74765 8.12109 9.71484 8.06719 9.65624L5.9625 7.37812L6.85547 6.41015L8.72578 8.4328H9V7.3078C9 6.80624 9.62578 6.55546 9.97969 6.90937L11.8359 8.57343C12.0539 8.79843 12.0539 9.15468 11.8359 9.37499ZM0.28125 3.56249H2.43281L4.30313 5.58515L5.19609 4.61718L3.09141 2.33905C3.0375 2.28046 2.9625 2.24765 2.88516 2.24765H0.28125C0.126563 2.24765 0 2.37421 0 2.5289V3.2789C0 3.43593 0.126563 3.56249 0.28125 3.56249ZM11.8359 2.58046L9.97969 0.916398C9.62578 0.562492 9 0.813273 9 1.31249V2.24999H8.27344C8.19609 2.24999 8.12109 2.2828 8.06719 2.3414L2.43281 8.43749H0.28125C0.126563 8.43749 0 8.56405 0 8.71874V9.46874C0 9.62343 0.126563 9.74999 0.28125 9.74999H2.88281C2.96016 9.74999 3.03516 9.71718 3.08906 9.65859L8.72344 3.56249H9V4.68749C9 5.18905 9.62578 5.43984 9.97969 5.08593L11.8359 3.37499C12.0539 3.15468 12.0539 2.79843 11.8359 2.58046Z"/>
-    </svg>
-  </span>
-</button>
